test_divergence2 vs test_divergence: the second version is as fast as fortran after compilation (0.1s to compile) while the first version takes ages to compile and is 100x slower than fortran. The magic is made by lax: lax.fori_loop

+++ data_types.py
@@ -21,53 +21,53 @@ class Mesh:
     nod2D: int = 0
     ocean_area: float = 0.0
     ocean_areawithcav: float = 0.0
-    coord_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    geo_coord_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
+    coord_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    geo_coord_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
     edge2D: int = 0
     edge2D_in: int = 0
-    elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    edges: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    edge_tri: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    elem_edges: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    elem_area: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    edge_dxdy: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    edge_cross_dxdy: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    elem_cos: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    metric_factor: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    elem_neighbors: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    nod_in_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.int32))
-    x_corners: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    y_corners: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    nod_in_elem2D_num: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    depth: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    gradient_vec: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    gradient_sca: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
+    elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    edges: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    edge_tri: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    elem_edges: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    elem_area: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    edge_dxdy: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    edge_cross_dxdy: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    elem_cos: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    metric_factor: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    elem_neighbors: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    nod_in_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    x_corners: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    y_corners: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    nod_in_elem2D_num: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    depth: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    gradient_vec: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    gradient_sca: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
     nl: int = 0
-    zbar: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    Z: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    elem_depth: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    ulevels: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    ulevels_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    nlevels: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    area: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0), dtype=jnp.float32))
-    mesh_resolution: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    cavity_flag_n: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    cavity_flag_e: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    zbar: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    Z: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    elem_depth: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    ulevels: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    ulevels_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    nlevels: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    area: jnp.ndarray = field(default_factory=lambda: jnp.zeros((0, 0)))
+    mesh_resolution: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    cavity_flag_n: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    cavity_flag_e: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     ssh_stiff: SparseMatrix = field(default_factory=SparseMatrix)
-    coriolis: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
-    coriolis_node: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.float32))
+    coriolis: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    coriolis_node: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
 
 # CommunicationStruct class (com_struct equivalent)
 @dataclass
 class CommunicationStruct:
     rPEnum: int = 0
-    rPE: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    rptr: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    rlist: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    rPE: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    rptr: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    rlist: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     sPEnum: int = 0
-    sPE: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    sptr: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    slist: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    sPE: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    sptr: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    slist: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
 
 # Partitioning class (T_PARTIT equivalent)
 @dataclass
@@ -79,22 +79,22 @@ class Partitioning:
     com_elem2D_full: CommunicationStruct = field(default_factory=CommunicationStruct)
     npes: int = 0
     mype: int = 0
-    part: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    part: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     myDim_nod2D: int = 0
     eDim_nod2D: int = 0
-    myList_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    myList_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     myDim_elem2D: int = 0
     eDim_elem2D: int = 0
     eXDim_elem2D: int = 0
-    myList_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    myList_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     myDim_edge2D: int = 0
     eDim_edge2D: int = 0
-    myList_edge2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    myList_edge2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
     pe_status: int = 0
     MPI_COMM_FESOM: int = 0
     MPI_COMM_WORLD: int = 0
-    remPtr_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    remList_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    remPtr_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
-    remList_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0, dtype=jnp.int32))
+    remPtr_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    remList_nod2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    remPtr_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
+    remList_elem2D: jnp.ndarray = field(default_factory=lambda: jnp.zeros(0))
